Rename package name to ubiconfig
mix up repo name and package name
