+++ README.md
@@ -1,7 +1,7 @@
-python-ubi-config
-==================
+ubiconfig
+=========
 
-A Python library for reading UBI configurations
+A Python library for accessing Universal Base Image configurations
 
 [![Build Status](https://travis-ci.org/release-engineering/ubi-config.svg?branch=master)](https://travis-ci.org/release-engineering/ubi-config)
 [![Coverage Status](https://coveralls.io/repos/github/release-engineering/ubi-config/badge.svg?branch=master)](https://coveralls.io/github/release-engineering/ubi-config?branch=master)
@@ -26,7 +26,7 @@ When there is `DEFAULT_UBI_REPO` set, user can load the config by passing the co
 name to `get_loader().load()`
 
 ```python
-from ubi_config import get_loader
+from ubiconfig import get_loader
 
 config = get_loader().load('rhel-8-for-x86_64-appstream')
 # config has been validated and is now a Python object with relevant properties
@@ -36,14 +36,14 @@ print package_whitelist
 Or, get all config files from the repo:
 ```python
 
-from ubi_config import get_loader
+from ubiconfig import get_loader
 
 configs = get_loader().load_all()
 # returns a list of UbiConfig objects
 ```
 Or, user can also load the config from local file:
 ```python
-from ubi_config import get_loader
+from ubiconfig import get_loader
 
 config = get_loader(local=True).load('/path/to/rhel-8-for-x86_64-appstream.yaml')
 ```
