+++ tox.ini
@@ -16,7 +16,7 @@ deps=
 commands=
 	flake8 \
 	  --max-complexity 10
-	sh -c 'pylint ubi_config; test $(( $? & (1|2|4|32) )) = 0'
+	sh -c 'pylint ubiconfig; test $(( $? & (1|2|4|32) )) = 0'
 
 [testenv:cov]
 deps=
@@ -24,7 +24,7 @@ deps=
 	pytest-cov
 usedevelop=true
 commands=
-	pytest --cov-report=html --cov=ubi_config {posargs}
+	pytest --cov-report=html --cov=ubiconfig {posargs}
 
 [testenv:cov-travis]
 passenv = TRAVIS TRAVIS_*
@@ -34,7 +34,7 @@ deps=
 	coveralls
 usedevelop=true
 commands=
-	pytest --cov=ubi_config {posargs}
+	pytest --cov=ubiconfig {posargs}
 	coveralls
 
 [testenv:docs]
