--- ubi_config/__init__.py
@@ -1,3 +1,0 @@
-from ubi_config.ubi import get_loader, Loader, UbiConfig
-
-__all__ = ['get_loader', 'Loader', 'UbiConfig']
