+++ ubiconfig/__init__.py
@@ -1,0 +1,3 @@
+from ubiconfig.ubi import get_loader, Loader, UbiConfig
+
+__all__ = ['get_loader', 'Loader', 'UbiConfig']
